Fixed the issue alls well
The issue was a problem of incorrect quotations that was expected but unfortunate in design. Once fixed the first four arrays were replaced with the new better array and grew to 814 in length. SOme the effacts are not new but as going through them al was deemed impossible I will allow the slip